+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -6,6 +6,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.Optional;
 import nextstep.mvc.controller.HandlerMappingRegistry;
+import nextstep.mvc.controller.adapter.HandlerAdapter;
+import nextstep.mvc.controller.adapter.HandlerAdapterRegistry;
 import nextstep.mvc.view.ModelAndView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
