+++ mvc/src/main/java/nextstep/mvc/controller/adapter/HandlerAdapter.java
@@ -1,4 +1,4 @@
-package nextstep.mvc;
+package nextstep.mvc.controller.adapter;
 
 import nextstep.mvc.view.ModelAndView;
